# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/stores/{storeId}/staff": {
      "get": {
        "operationId": "StaffController_listStaff",
        "summary": "List staff members",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of staff"
          }
        },
        "tags": [
          "Staff"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "StaffController_inviteStaff",
        "summary": "Invite a staff member",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Staff member invited"
          },
          "409": {
            "description": "Email already exists"
          }
        },
        "tags": [
          "Staff"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/stores/{storeId}/staff/{id}": {
      "get": {
        "operationId": "StaffController_getStaff",
        "summary": "Get staff member details",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Staff member details"
          },
          "404": {
            "description": "Staff member not found"
          }
        },
        "tags": [
          "Staff"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "StaffController_updateStaff",
        "summary": "Update staff role/permissions",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Staff member updated"
          },
          "404": {
            "description": "Staff member not found"
          }
        },
        "tags": [
          "Staff"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "StaffController_removeStaff",
        "summary": "Remove staff member",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Staff member removed"
          },
          "404": {
            "description": "Staff member not found"
          }
        },
        "tags": [
          "Staff"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/stores/{storeId}/warehouses": {
      "get": {
        "operationId": "WarehousesController_listWarehouses",
        "summary": "List warehouses",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of warehouses"
          }
        },
        "tags": [
          "Inventory"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "WarehousesController_createWarehouse",
        "summary": "Create a warehouse",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Warehouse created"
          }
        },
        "tags": [
          "Inventory"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/stores/{storeId}/warehouses/{warehouseId}": {
      "get": {
        "operationId": "WarehousesController_getWarehouse",
        "summary": "Get warehouse details",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "warehouseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Warehouse details"
          },
          "404": {
            "description": "Warehouse not found"
          }
        },
        "tags": [
          "Inventory"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "WarehousesController_updateWarehouse",
        "summary": "Update warehouse",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "warehouseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Warehouse updated"
          },
          "404": {
            "description": "Warehouse not found"
          }
        },
        "tags": [
          "Inventory"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "WarehousesController_deleteWarehouse",
        "summary": "Delete warehouse (must have no inventory)",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "warehouseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Warehouse deleted"
          },
          "409": {
            "description": "Warehouse has inventory records"
          }
        },
        "tags": [
          "Inventory"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/stores/{storeId}/inventory/warehouse/{warehouseId}": {
      "get": {
        "operationId": "InventoryController_listByWarehouse",
        "summary": "List inventory for a warehouse",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "warehouseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Inventory"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "ShopBuilder API",
    "description": "ShopBuilder — Shopify Clone for Local Kazakh Merchants. E-commerce platform API.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Paste JWT access token (without \"Bearer \" prefix)"
      }
    },
...
  }
}